make the crate #[no_std]
The tests do use std::sync::Arc, so move them into an integration tests
file, so that they still build with std.
